Split navItem into several components
