+++ portal/app/[locale]/_components/navbar/_components/externalLink.tsx
@@ -1,0 +1,57 @@
+import { useUmami } from 'app/analyticsEvents'
+import { ExternalLink as AnchorTag } from 'components/externalLink'
+import { ArrowDownLeftIcon } from 'components/icons/arrowDownLeftIcon'
+import { Link } from 'components/link'
+import { useNetworkType } from 'hooks/useNetworkType'
+import { ComponentProps } from 'react'
+
+import {
+  IconContainer,
+  ItemContainer,
+  ItemText,
+  NavItemProps,
+  Row,
+} from './navItem'
+
+type ItemLinkProps = NavItemProps &
+  Required<Pick<ComponentProps<typeof Link>, 'href'>>
+
+const ExternalLinkUI = ({
+  href,
+  icon,
+  onClick,
+  text,
+}: Omit<ItemLinkProps, 'href'> &
+  Pick<ComponentProps<'a'>, 'href' | 'onClick'>) => (
+  <ItemContainer>
+    <AnchorTag href={href} onClick={onClick}>
+      <Row>
+        {icon && <IconContainer>{icon}</IconContainer>}
+        <ItemText text={text} />
+        <div className="ml-auto hidden size-4 items-center group-hover/item:flex">
+          <ArrowDownLeftIcon />
+        </div>
+      </Row>
+    </AnchorTag>
+  </ItemContainer>
+)
+
+export const ExternalLink = function ({
+  event,
+  href,
+  icon,
+  text,
+}: Omit<ItemLinkProps, 'href'> & Pick<ComponentProps<'a'>, 'href'>) {
+  const [networkType] = useNetworkType()
+  const { track } = useUmami()
+  const addTracking = () =>
+    track ? () => track(event, { chain: networkType }) : undefined
+  return (
+    <ExternalLinkUI
+      href={href}
+      icon={icon}
+      onClick={addTracking()}
+      text={text}
+    />
+  )
+}
