+++ portal/app/[locale]/_components/navbar/index.tsx
@@ -18,7 +18,8 @@ import { GetStarted } from './_components/getStarted'
 import { Help } from './_components/help'
 import { HemiExplorerLink } from './_components/hemiExplorerLink'
 import { HemiLogoFull } from './_components/hemiLogo'
-import { ItemLink, NetworkSwitch } from './_components/navItem'
+import { ItemLink } from './_components/itemLink'
+import { NetworkSwitch } from './_components/navItem'
 import { SocialLinks } from './_components/socialLinks'
 import { Tvl } from './_components/tvl'
